Arch Linuxかどうかの処理を追加

+++ .bashrc
@@ -23,7 +23,7 @@ HISTFILESIZE=2000
 alias ls='ls --color=auto'
 alias lr='ls -R'                    # recursive ls
 alias ll='ls -l'
-alias la='ll -la'
+alias la='ll -lA'
 alias lx='ll -BX'                   # sort by extension
 alias lz='ll -rS'                   # sort by size
 alias lt='ll -rt'                   # sort by date
@@ -65,21 +65,24 @@ alias hist='history | grep $1'      # requires an argument
 alias openports='netstat --all --numeric --programs --inet --inet6'
 
 
-# pacman aliases (if applicable, replace 'pacman' with 'yaourt'/'pacaur'/whatever)
-alias pac="pacman -S"      # default action     - install one or more packages
-alias pacu="pacman -Syu"   # '[u]pdate'         - upgrade all packages to their newest version
-alias pacs="pacman -Ss"    # '[s]earch'         - search for a package using one or more keywords
-alias paci="pacman -Si"    # '[i]nfo'           - show information about a package
-alias pacr="pacman -R"     # '[r]emove'         - uninstall one or more packages
-alias pacl="pacman -Sl"    # '[l]ist'           - list all packages of a repository
-alias pacll="pacman -Qqm"  # '[l]ist [l]ocal'   - list all packages which were locally installed (e.g. AUR packages)
-alias paclo="pacman -Qdt"  # '[l]ist [o]rphans' - list all packages which are orphaned
-alias paco="pacman -Qo"    # '[o]wner'          - determine which package owns a given file
-alias pacf="pacman -Ql"    # '[f]iles'          - list all files installed by a given package
-alias pacc="pacman -Sc"    # '[c]lean cache'    - delete all not currently installed package files
-alias pacm="makepkg -fci"  # '[m]ake'           - make package from PKGBUILD file in current directory
 
-# @see https://wiki.archlinux.org/index.php/Bashrc#Aliases
+# pacman aliases (if applicable, replace 'pacman' with 'yaourt'/'pacaur'/whatever)
+if [[ `uname -r` =~ ARCH$ ]]; then
+    alias pac="pacman -S"      # default action     - install one or more packages
+    alias pacu="pacman -Syu"   # '[u]pdate'         - upgrade all packages to their newest version
+    alias pacs="pacman -Ss"    # '[s]earch'         - search for a package using one or more keywords
+    alias paci="pacman -Si"    # '[i]nfo'           - show information about a package
+    alias pacr="pacman -R"     # '[r]emove'         - uninstall one or more packages
+    alias pacl="pacman -Sl"    # '[l]ist'           - list all packages of a repository
+    alias pacll="pacman -Qqm"  # '[l]ist [l]ocal'   - list all packages which were locally installed (e.g. AUR packages)
+    alias paclo="pacman -Qdt"  # '[l]ist [o]rphans' - list all packages which are orphaned
+    alias paco="pacman -Qo"    # '[o]wner'          - determine which package owns a given file
+    alias pacf="pacman -Ql"    # '[f]iles'          - list all files installed by a given package
+    alias pacc="pacman -Sc"    # '[c]lean cache'    - delete all not currently installed package files
+    alias pacm="makepkg -fci"  # '[m]ake'           - make package from PKGBUILD file in current directory
+    # @see https://wiki.archlinux.org/index.php/Bashrc#Aliases
+    alias yaupdate="yaourt -Syua --devel"
+fi
 
 extract() {
     local c e i
@@ -116,3 +119,6 @@ extract() {
 
     return $e
 }
+
+PATH=$PATH:~/.cabal/bin
+export PATH
